Fix/oop layout and diagnostics

+++ packages/spectra-api/src/trace.rs
@@ -7,10 +7,10 @@ unsafe fn string(value: i64) -> Option<String> {
     if value == 0 {
         return None;
     }
-    let ptr = value as *const i64;
+    let ptr = value as *const u8;
     let mut bytes = Vec::new();
     for index in 0..4096 {
-        let byte = *ptr.add(index) as u8;
+        let byte = *ptr.add(index);
         if byte == 0 {
             return String::from_utf8(bytes).ok();
         }
